+++ extensions/music/src/main/java/app/revanced/extension/music/patches/spoof/SpoofVideoStreamsPatch.java
@@ -1,5 +1,6 @@
 package app.revanced.extension.music.patches.spoof;
 
+import static app.revanced.extension.music.settings.Settings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE;
 import static app.revanced.extension.shared.spoof.ClientType.ANDROID_VR_1_43_32;
 import static app.revanced.extension.shared.spoof.ClientType.ANDROID_VR_1_61_48;
 import static app.revanced.extension.shared.spoof.ClientType.VISIONOS;
@@ -7,7 +8,6 @@ import static app.revanced.extension.shared.spoof.ClientType.VISIONOS;
 import java.util.List;
 
 import app.revanced.extension.shared.spoof.ClientType;
-import app.revanced.extension.shared.spoof.requests.StreamingDataRequest;
 
 @SuppressWarnings("unused")
 public class SpoofVideoStreamsPatch {
@@ -22,6 +22,7 @@ public class SpoofVideoStreamsPatch {
                 VISIONOS
         );
 
-        StreamingDataRequest.setClientOrderToUse(availableClients, ANDROID_VR_1_43_32);
+        app.revanced.extension.shared.spoof.SpoofVideoStreamsPatch.setClientsToUse(
+                availableClients, SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get());
     }
 }