+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/TrieSearch.java
@@ -1,6 +1,5 @@
-package app.revanced.extension.youtube;
+package app.revanced.extension.shared;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
@@ -57,11 +56,13 @@ public abstract class TrieSearch<T> {
             if (searchTextLength - searchTextIndex < patternLength - patternStartIndex) {
                 return false; // Remaining search text is shorter than the remaining leaf pattern and they cannot match.
             }
+
             for (int i = searchTextIndex, j = patternStartIndex; j < patternLength; i++, j++) {
                 if (enclosingNode.getCharValue(searchText, i) != enclosingNode.getCharValue(pattern, j)) {
                     return false;
                 }
             }
+
             return callback == null || callback.patternMatched(searchText,
                     searchTextIndex - patternStartIndex, patternLength, callbackParameter);
         }
@@ -136,7 +137,7 @@ public abstract class TrieSearch<T> {
          * @param patternLength Length of the pattern.
          * @param callback      Callback, where a value of NULL indicates to always accept a pattern match.
          */
-        private void addPattern(@NonNull T pattern, int patternIndex, int patternLength,
+        private void addPattern(T pattern, int patternIndex, int patternLength,
                                 @Nullable TriePatternMatchedCallback<T> callback) {
             if (patternIndex == patternLength) { // Reached the end of the pattern.
                 if (endOfPatternCallback == null) {
@@ -145,6 +146,7 @@ public abstract class TrieSearch<T> {
                 endOfPatternCallback.add(callback);
                 return;
             }
+
             if (leaf != null) {
                 // Reached end of the graph and a leaf exist.
                 // Recursively call back into this method and push the existing leaf down 1 level.
@@ -159,6 +161,7 @@ public abstract class TrieSearch<T> {
                 leaf = new TrieCompressedPath<>(pattern, patternIndex, patternLength, callback);
                 return;
             }
+
             final char character = getCharValue(pattern, patternIndex);
             final int arrayIndex = hashIndexForTableSize(children.length, character);
             TrieNode<T> child = children[arrayIndex];
@@ -183,6 +186,7 @@ public abstract class TrieSearch<T> {
                 //noinspection unchecked
                 TrieNode<T>[] replacement = new TrieNode[replacementArraySize];
                 addNodeToArray(replacement, child);
+
                 boolean collision = false;
                 for (TrieNode<T> existingChild : children) {
                     if (existingChild != null) {
@@ -195,6 +199,7 @@ public abstract class TrieSearch<T> {
                 if (collision) {
                     continue;
                 }
+
                 children = replacement;
                 return;
             }
@@ -234,6 +239,7 @@ public abstract class TrieSearch<T> {
                 if (leaf != null && leaf.matches(startNode, searchText, searchTextEndIndex, searchTextIndex, callbackParameter)) {
                     return true; // Leaf exists and it matched the search text.
                 }
+
                 List<TriePatternMatchedCallback<T>> endOfPatternCallback = node.endOfPatternCallback;
                 if (endOfPatternCallback != null) {
                     final int matchStartIndex = searchTextIndex - currentMatchLength;
@@ -246,6 +252,7 @@ public abstract class TrieSearch<T> {
                         }
                     }
                 }
+
                 TrieNode<T>[] children = node.children;
                 if (children == null) {
                     return false; // Reached a graph end point and there's no further patterns to search.
@@ -278,9 +285,11 @@ public abstract class TrieSearch<T> {
             if (leaf != null) {
                 numberOfPointers += 4; // Number of fields in leaf node.
             }
+
             if (endOfPatternCallback != null) {
                 numberOfPointers += endOfPatternCallback.size();
             }
+
             if (children != null) {
                 numberOfPointers += children.length;
                 for (TrieNode<T> child : children) {
@@ -308,13 +317,13 @@ public abstract class TrieSearch<T> {
     private final List<T> patterns = new ArrayList<>();
 
     @SafeVarargs
-    TrieSearch(@NonNull TrieNode<T> root, @NonNull T... patterns) {
+    TrieSearch(TrieNode<T> root, T... patterns) {
         this.root = Objects.requireNonNull(root);
         addPatterns(patterns);
     }
 
     @SafeVarargs
-    public final void addPatterns(@NonNull T... patterns) {
+    public final void addPatterns(T... patterns) {
         for (T pattern : patterns) {
             addPattern(pattern);
         }
@@ -325,7 +334,7 @@ public abstract class TrieSearch<T> {
      *
      * @param pattern Pattern to add. Calling this with a zero length pattern does nothing.
      */
-    public void addPattern(@NonNull T pattern) {
+    public void addPattern(T pattern) {
         addPattern(pattern, root.getTextLength(pattern), null);
     }
 
@@ -333,31 +342,31 @@ public abstract class TrieSearch<T> {
      * @param pattern  Pattern to add. Calling this with a zero length pattern does nothing.
      * @param callback Callback to determine if searching should halt when a match is found.
      */
-    public void addPattern(@NonNull T pattern, @NonNull TriePatternMatchedCallback<T> callback) {
+    public void addPattern(T pattern, TriePatternMatchedCallback<T> callback) {
         addPattern(pattern, root.getTextLength(pattern), Objects.requireNonNull(callback));
     }
 
-    void addPattern(@NonNull T pattern, int patternLength, @Nullable TriePatternMatchedCallback<T> callback) {
+    void addPattern(T pattern, int patternLength, @Nullable TriePatternMatchedCallback<T> callback) {
         if (patternLength == 0) return; // Nothing to match
 
         patterns.add(pattern);
         root.addPattern(pattern, 0, patternLength, callback);
     }
 
-    public final boolean matches(@NonNull T textToSearch) {
+    public final boolean matches(T textToSearch) {
         return matches(textToSearch, 0);
     }
 
-    public boolean matches(@NonNull T textToSearch, @NonNull Object callbackParameter) {
+    public boolean matches(T textToSearch, Object callbackParameter) {
         return matches(textToSearch, 0, root.getTextLength(textToSearch),
                 Objects.requireNonNull(callbackParameter));
     }
 
-    public boolean matches(@NonNull T textToSearch, int startIndex) {
+    public boolean matches(T textToSearch, int startIndex) {
         return matches(textToSearch, startIndex, root.getTextLength(textToSearch));
     }
 
-    public final boolean matches(@NonNull T textToSearch, int startIndex, int endIndex) {
+    public final boolean matches(T textToSearch, int startIndex, int endIndex) {
         return matches(textToSearch, startIndex, endIndex, null);
     }
 
@@ -370,11 +379,11 @@ public abstract class TrieSearch<T> {
      * @param callbackParameter Optional parameter passed to the callbacks.
      * @return If any pattern matched, and it's callback halted searching.
      */
-    public boolean matches(@NonNull T textToSearch, int startIndex, int endIndex, @Nullable Object callbackParameter) {
+    public boolean matches(T textToSearch, int startIndex, int endIndex, @Nullable Object callbackParameter) {
         return matches(textToSearch, root.getTextLength(textToSearch), startIndex, endIndex, callbackParameter);
     }
 
-    private boolean matches(@NonNull T textToSearch, int textToSearchLength, int startIndex, int endIndex,
+    private boolean matches(T textToSearch, int textToSearchLength, int startIndex, int endIndex,
                             @Nullable Object callbackParameter) {
         if (endIndex > textToSearchLength) {
             throw new IllegalArgumentException("endIndex: " + endIndex