+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/spoof/requests/StreamingDataRequest.java
@@ -1,5 +1,6 @@
 package app.revanced.extension.shared.spoof.requests;
 
+import static app.revanced.extension.shared.ByteTrieSearch.convertStringsToBytes;
 import static app.revanced.extension.shared.spoof.requests.PlayerRoutes.GET_STREAMING_DATA;
 
 import androidx.annotation.NonNull;
@@ -23,6 +24,7 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import app.revanced.extension.shared.ByteTrieSearch;
 import app.revanced.extension.shared.Logger;
 import app.revanced.extension.shared.Utils;
 import app.revanced.extension.shared.settings.BaseSettings;
@@ -89,6 +91,16 @@ public class StreamingDataRequest {
      */
     private static final Map<String, StreamingDataRequest> cache = Collections.synchronizedMap(
             Utils.createSizeRestrictedMap(50));
+
+    /**
+     * Strings found in the response if the video is a livestream.
+     */
+    private static final ByteTrieSearch liveStreamBufferSearch = new ByteTrieSearch(
+            convertStringsToBytes(
+                    "yt_live_broadcast",
+                    "yt_premiere_broadcast"
+            )
+    );
 
     private static volatile ClientType lastSpoofedClientType;
 
@@ -157,7 +169,7 @@ public class StreamingDataRequest {
                 }
             }
 
-            if (!authHeadersIncludes && clientType.requiresAuth) {
+            if (!authHeadersIncludes && clientType.useAuth) {
                 Logger.printDebug(() -> "Skipping client since user is not logged in: " + clientType
                         + " videoId: " + videoId);
                 return null;
@@ -218,9 +230,13 @@ public class StreamingDataRequest {
                             while ((bytesRead = inputStream.read(buffer)) >= 0) {
                                 baos.write(buffer, 0, bytesRead);
                             }
-                            lastSpoofedClientType = clientType;
+                            if (clientType == ClientType.ANDROID_CREATOR && liveStreamBufferSearch.matches(buffer)) {
+                                Logger.printDebug(() -> "Skipping Android Studio as video is a livestream: " + videoId);
+                            } else {
+                                lastSpoofedClientType = clientType;
 
-                            return ByteBuffer.wrap(baos.toByteArray());
+                                return ByteBuffer.wrap(baos.toByteArray());
+                            }
                         }
                     }
                 } catch (IOException ex) {