+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/ReturnYouTubeDislikeFilter.java
@@ -8,8 +8,8 @@ import java.util.LinkedHashSet;
 import java.util.Map;
 
 import app.revanced.extension.shared.Logger;
+import app.revanced.extension.shared.TrieSearch;
 import app.revanced.extension.shared.Utils;
-import app.revanced.extension.youtube.TrieSearch;
 import app.revanced.extension.youtube.patches.ReturnYouTubeDislikePatch;
 import app.revanced.extension.youtube.patches.VideoInformation;
 import app.revanced.extension.youtube.settings.Settings;