+++ patches/src/main/kotlin/app/revanced/patches/shared/misc/spoof/Fingerprints.kt
@@ -146,7 +146,6 @@ internal val nerdsStatsVideoFormatBuilderFingerprint by fingerprint {
 }
 
 internal val patchIncludedExtensionMethodFingerprint by fingerprint {
-    accessFlags(AccessFlags.PRIVATE, AccessFlags.STATIC)
     returns("Z")
     parameters()
     custom { method, classDef ->