+++ patches/src/main/kotlin/app/revanced/patches/viber/ads/Fingerprints.kt
@@ -2,12 +2,6 @@ package app.revanced.patches.viber.ads
 
 import app.revanced.patcher.fingerprint
 
-internal val adsFreeFingerprint by fingerprint {
-    returns("I")
-    parameters()
-    custom { method, classDef ->
-        classDef.type.contains("com/viber/voip/feature/viberplus") &&
-        classDef.superclass?.contains("com/viber/voip/core/feature") == true &&  // Must extend com.viber.voip.core.feature.?
-        classDef.methods.count() == 1
-    }
+internal val findAdStringFingerprint by fingerprint {
+    strings("viber_plus_debug_ads_free_flag")
 }