+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/buttons/action/HideButtonsPatch.kt
@@ -55,7 +55,7 @@ val hideButtonsPatch = resourcePatch(
                 SwitchPreference("revanced_hide_remix_button"),
                 SwitchPreference("revanced_hide_report_button"),
                 SwitchPreference("revanced_hide_share_button"),
-                SwitchPreference("revanced_hide_stop_ads_button"),
+                SwitchPreference("revanced_hide_shop_button"),SwitchPreference("revanced_hide_stop_ads_button"),
                 SwitchPreference("revanced_hide_thanks_button"),
             ))
         }